Fix cannot delete rows bug

+++ backend/foodbank_southlondon/api/utils.py
@@ -64,7 +64,8 @@ def overwrite_rows(spreadsheet_id: str, rows: List, index: int = 0) -> None:
     flask.current_app.logger.debug(f"Overwriting all rows with {new_row_count} new rows in {sheet.spreadsheet.title} ({sheet.url}) ...")
     existing_row_count = sheet.row_count
     sheet.update(rows, value_input_option="USER_ENTERED")
-    sheet.delete_rows(new_row_count + 1, existing_row_count)
+    if existing_row_count > new_row_count:
+        sheet.delete_rows(new_row_count + 1, existing_row_count)
 
 
 def paginate(*sort_by: str, ascending: bool = True) -> Callable:
